Adjusting namespaces and adding documentation
