+++ src/With/Collections/ComparerFromFunc.cs
@@ -5,7 +5,7 @@ namespace With.Collections
     /// <summary>
     /// Class that holds methods related to comparers
     /// </summary>
-    internal partial class Comparer
+    internal class Comparer
     {
         /// <summary>
         /// Create a <see cref="System.Collections.Generic.IComparer{T}"/> from a comparison Func
@@ -13,32 +13,53 @@ namespace With.Collections
         /// <typeparam name="T"></typeparam>
         /// <param name="compare"></param>
         /// <returns></returns>
-        public static ComparerFromFunc<T> Create<T>(Func<T, T, int> compare)
+        public static IComparer<T> Create<T>(Func<T, T, int> compare)
         {
             return new ComparerFromFunc<T>(compare);
         }
-    }
-    /// <summary>
-    /// An adapter class to let a compare func act as a <see cref="System.Collections.Generic.IComparer{T}"/>
-    /// </summary>
-    /// <typeparam name="T"></typeparam>
-    internal class ComparerFromFunc<T> : IComparer<T>
-    {
-        private readonly Func<T, T, int> _compare;
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="compare"></param>
-        public ComparerFromFunc(Func<T, T, int> compare)
+        public static IComparer<T> Create<T, TComparable>(Func<T, TComparable> select)
+            where TComparable : IComparable
         {
-            _compare = compare;
+            return new ComparerFromSelect<T, TComparable>(select);
+        }
+        private class ComparerFromSelect<T, TComparable> : IComparer<T> where TComparable : IComparable
+        {
+            private readonly Func<T, TComparable> _select;
+            private readonly IComparer<TComparable> _comparer = Comparer<TComparable>.Default;
+
+            public ComparerFromSelect(Func<T, TComparable> select)
+            {
+                _select = select;
+            }
+
+            public int Compare(T x, T y)
+            {
+                return _comparer.Compare(_select(x), _select(y));
+            }
         }
         /// <summary>
-        /// Compares two objects and returns a value indicating the ordering between the two objects.
+        /// An adapter class to let a compare func act as a <see cref="System.Collections.Generic.IComparer{T}"/>
         /// </summary>
-        public int Compare(T x, T y)
+        /// <typeparam name="T"></typeparam>
+        private class ComparerFromFunc<T> : IComparer<T>
         {
-            return _compare(x, y);
+            private readonly Func<T, T, int> _compare;
+            /// <summary>
+            /// 
+            /// </summary>
+            /// <param name="compare"></param>
+            public ComparerFromFunc(Func<T, T, int> compare)
+            {
+                _compare = compare;
+            }
+            /// <summary>
+            /// Compares two objects and returns a value indicating the ordering between the two objects.
+            /// </summary>
+            public int Compare(T x, T y)
+            {
+                return _compare(x, y);
+            }
         }
     }
+   
 }
