+++ src/With/Ranges/IRange.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using With.Collections;
 
 namespace With.Ranges
 {
