+++ src/With/RegularExpressions/GrepExtensions.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using With.Linq;
+using With.Collections;
 
 namespace With.RegularExpressions
 {
