+++ tests/Tests/Collections/FlattenTests.cs
@@ -1,9 +1,9 @@
 using System.Collections;
-using Xunit;
-using With.Linq;
 using System.Linq;
+using With.Collections;
+using Xunit;
 
-namespace With.Tests.Linq
+namespace Tests.Collections
 {
     public class FlattenTests
     {
